New search functionality
The search system now looks for entries which contain all of the terms entered into the search bar. If the terms are enclosed in quotes, the terms will be treated as a phrase. Only entries containing the entire phrase will be matched. Search results can now be ordered by folio ID or relevance. Also, the search system no longer attempts to stem text. This means that it will only match terms that are pluralized and conjugated the same as the search term. Lastly, "stop words" have been eliminated, so common words will now match (the, or, etc.)